+++ lm-api/CHANGELOG.md
@@ -3,6 +3,7 @@
 This file keeps track of all notable changes to `License Manager API`.
 
 ## Unreleased
+* Expanded permission sets from view/edit to create/read/update/delete [PENG-2160]
 
 ## 3.1.0 -- 2024-01-24
 * Update constraints to limit int fields to 2**31-1 [PENG-1438]